Renamed Python and Node Files to Snake Case
Follows .js and .py file name conventions.
